run klasse implementeret
run klasse men stortset samme som før
