+++ run.h
@@ -1,0 +1,19 @@
+#ifndef RUN_H
+#define RUN_H
+#include "sound_generator.h"
+#include <ncurses.h>
+
+enum direction
+{
+    FORWARD, BACKWARDS, LEFT, RIGHT
+};
+
+class Run : SoundGenerator
+{
+public:
+    Run();
+    void run();
+    void Movement(direction d);
+};
+
+#endif // RUN_H
