+++ sound_generator.h
@@ -5,24 +5,16 @@
 #include <vector>
 #include <math.h>
 #include <unistd.h>
-#include <ncurses.h>
-
-enum direction
-{
-    FORWARD, BACKWARDS, LEFT, RIGHT
-};
 
 class SoundGenerator
 {
 public:
     SoundGenerator();
-    void Movement(direction d);
     void PlaySingle(double f1, double f2);
-    void Run();
-private:
     void PlayLoop(double f1, double f2);
-    short SineWave(double time, double freq1, double freq2, double amp);
     sf::Sound lsound;
+private:
+    short SineWave(double time, double freq1, double freq2, double amp);
     sf::SoundBuffer lbuffer;
 };
